+++ application/models/login_model.php
@@ -318,7 +318,7 @@ class LoginModel
         // set the rememberme-cookie to ten years ago (3600sec * 365 days * 10).
         // that's obviously the best practice to kill a cookie via php
         // @see http://stackoverflow.com/a/686166/1114320
-        setcookie('rememberme', false, time() - (3600 * 3650), '/');
+        setcookie('rememberme', false, time() - (3600 * 3650), '/', COOKIE_DOMAIN);
     }
 
     /**
@@ -355,7 +355,7 @@ class LoginModel
         }
 
         // clean the input
-        $user_name = substr(htmlentities($_POST['user_name'], ENT_QUOTES), 0, 64);
+        $user_name = substr(strip_tags($_POST['user_name']), 0, 64);
 
         // check if new username already exists
         $query = $this->db->prepare("SELECT user_id FROM users WHERE user_name = :user_name");
@@ -413,7 +413,7 @@ class LoginModel
         }
 
         // cleaning and write new email to database
-        $user_email = substr(htmlentities($_POST['user_email'], ENT_QUOTES), 0, 64);
+        $user_email = substr(strip_tags($_POST['user_email']), 0, 64);
         $query = $this->db->prepare("UPDATE users SET user_email = :user_email WHERE user_id = :user_id");
         $query->execute(array(':user_email' => $user_email, ':user_id' => $_SESSION['user_id']));
         $count =  $query->rowCount();
@@ -469,8 +469,8 @@ class LoginModel
             AND ($_POST['user_password_new'] === $_POST['user_password_repeat'])) {
 
             // clean the input
-            $user_name = htmlentities($_POST['user_name'], ENT_QUOTES);
-            $user_email = htmlentities($_POST['user_email'], ENT_QUOTES);
+            $user_name = strip_tags($_POST['user_name']);
+            $user_email = strip_tags($_POST['user_email']);
 
             // crypt the user's password with the PHP 5.5's password_hash() function, results in a 60 character
             // hash string. the PASSWORD_DEFAULT constant is defined by the PHP 5.5, or if you are using PHP 5.3/5.4,
@@ -691,7 +691,7 @@ class LoginModel
             return false;
         }
         // if input file too small
-        if ($image_proportions[0] < 100 OR $image_proportions[1] < 100) {
+        if ($image_proportions[0] < AVATAR_SIZE OR $image_proportions[1] < AVATAR_SIZE) {
             $_SESSION["feedback_negative"][] = FEEDBACK_AVATAR_UPLOAD_TOO_SMALL;
             return false;
         }
@@ -845,7 +845,7 @@ class LoginModel
         // generate random hash for email password reset verification (40 char string)
         $user_password_reset_hash = sha1(uniqid(mt_rand(), true));
         // clean user input
-        $user_name = htmlentities($_POST['user_name'], ENT_QUOTES);
+        $user_name = strip_tags($_POST['user_name']);
 
         // check if that username exists
         $query = $this->db->prepare("SELECT user_id, user_email FROM users
@@ -1233,11 +1233,15 @@ class LoginModel
 
         // check if a user with that username already exists in our database
         // note: Facebook's internal username is usually the person's full name plus a number (and dots between)
-        // TODO: if username is already taken, add number etc. to potential username and repeat this step
         if ($this->facebookUserNameExistsAlreadyInDatabase($facebook_user_data)) {
-            $_SESSION["feedback_negative"][] = FEEDBACK_FACEBOOK_USERNAME_ALREADY_EXISTS;
-            return false;
+        	$facebook_user_data["username"] = $this->generateUniqueUserNameFromExistingUserName($facebook_user_data["username"]);
+         if ($this->facebookUserNameExistsAlreadyInDatabase($facebook_user_data)) {
+        	//shouldn't get here if we managed to generate a unique name!
+        	$_SESSION["feedback_negative"][] = FEEDBACK_FACEBOOK_USERNAME_ALREADY_EXISTS;
+          return false;
+         }
         }
+
 
         // check if that email address already exists in our database
         if ($this->facebookUserEmailExistsAlreadyInDatabase($facebook_user_data)) {
@@ -1365,4 +1369,29 @@ class LoginModel
         // default return
         return false;
     }
+
+    /**
+     * Generate unique user_name from facebook-user's username appended with a number
+     * @param string $existing_name $facebook_user_data stuff from the facebook class
+     * @return string unique user_name not in database yet
+     */
+    public function generateUniqueUserNameFromExistingUserName($existing_name)
+    {
+    	//strip any dots, trailing numbers and white spaces
+        $existing_name = str_replace(".", "", $existing_name);
+        $existing_name = preg_replace('/\s*\d+$/', '', $existing_name);
+
+        // loop until we have a new username, adding an increasing number to the given string every time
+    	$n = 0;
+    	do {
+            $n = $n+1;
+            $new_username = $existing_name . $n;
+            $query = $this->db->prepare("SELECT user_id FROM users WHERE user_name = :name_with_number");
+            $query->execute(array(':name_with_number' => $new_username));
+    	 	 
+    	 } while ($query->rowCount() == 1);
+
+    	return $new_username;
+    }
+
 }